Refactor package name details
