+++ app/src/main/java/com/afoxplus/yalistoadmin/ui/graphs/HomeNavGraph.kt
@@ -11,12 +11,12 @@ import androidx.navigation.navArgument
 import com.afoxplus.yalistoadmin.commons.extensions.navigateSingleTopTo
 import com.afoxplus.yalistoadmin.commons.extensions.sharedViewModel
 import com.afoxplus.yalistoadmin.domain.entities.OrderNavType
-import com.afoxplus.yalistoadmin.ui.screens.order_status.OrderStatusScreen
 import com.afoxplus.yalistoadmin.ui.screens.home.BottomBarHomeRouter
 import com.afoxplus.yalistoadmin.ui.screens.home.orders.OrdersHomeScreen
-import com.afoxplus.yalistoadmin.ui.screens.orders.OrdersStatusViewModel
 import com.afoxplus.yalistoadmin.ui.screens.home.sales.SalesHomeScreen
+import com.afoxplus.yalistoadmin.ui.screens.orders.OrdersStatusViewModel
 import com.afoxplus.yalistoadmin.ui.screens.products.ProductScreen
+import com.afoxplus.yalistoadmin.ui.screens.status.OrderStatusScreen
 
 @Composable
 fun HomeNavGraph(navController: NavHostController, modifier: Modifier = Modifier) {
