+++ app/src/main/java/com/afoxplus/yalistoadmin/ui/screens/status/components/OrderAppetizersComponent.kt
@@ -1,4 +1,4 @@
-package com.afoxplus.yalistoadmin.ui.screens.order_status.components
+package com.afoxplus.yalistoadmin.ui.screens.status.components
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.fillMaxWidth
